requirements.txt
slackclient
allennlp
